requirements.txt
deepface==0.0.93
opencv-python==4.9.0.80
streamlit==1.35.0
numpy==1.26.4
Pillow==10.3.0
pandas==2.2.2
scikit-learn==1.4.2
tf-keras>=2.16.0
